requirements.txt
pandas==1.1.1
streamlit==0.66.0
Pillow==7.2.0
pandas-datareader==0.9.0